feat(service): configure port and targetPort sep
Separately configure port and targetPort for services.

This is because wg-easy hard-codes the iptables rule dport it uses to
51820 https://github.com/WeeJeWel/wg-easy/blob/master/src/config.js#L22

If you modified the port previously, it would not allow things to work
at all, thus port configurability wasn't working as intended.

This change allows you to configure them independently and allow you to
change `port` as intended while allowing you to keep the hard-coded
targetPort.